Create projects and users routes and clean main.js to abstract out some apis
